requirements.txt
click>=8.1
g4f>=0.3
tweepy>=4.14
google-api-python-client>=2.100
google-auth-oauthlib>=1.1
google-auth-httplib2>=0.1
requests>=2.31
beautifulsoup4>=4.12
apscheduler>=3.10
moviepy>=1.0.3
Pillow>=10.0
edge-tts>=6.1
openai>=1.30
